requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
pydantic==2.9.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.8
pytest==8.3.3
httpx==0.27.2
pytest-cov==5.0.0
flake8==7.1.1
black==24.8.0
